pyproject.toml
[project]
name = "hatch-messaging-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
]
